requirements.txt
# Django Core
Django==5.2.8
djangorestframework==3.16.1

# JWT Authentication
djangorestframework_simplejwt==5.5.1

# Database
psycopg2-binary==2.9.11
dj-database-url==3.0.1

# Configuration
django-environ==0.12.0
python-dotenv==1.2.1
python-decouple==3.8

# CORS
django-cors-headers==4.9.0

# Filtering
django-filter==25.2

# Documentation
drf-yasg==1.21.11

# Production Server
gunicorn==23.0.0
whitenoise==6.11.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

